requirements.txt
pyserial~=3.5
gpiozero~=2.0.1
paho-mqtt~=2.0.0
python-dotenv~=1.0.1
jinja2